feat: Update denols startup mode and links
This commit updates the startup mode for denols in the case where no
node-specific files are found. Instead of suppressing the startup, denols
is now started in Single file mode. If tsserver is running, it is stopped.

Additionally, the commit updates the links to the code snippets in the
dotfiles repository. The links now point to the updated lines of code in
the denols and vtsls server configurations, and the neoconf plugin.