feat: Add article about TypeScript exhaustiveness check
This commit introduces a new article that explains how to use the
`satisfies` keyword in TypeScript for exhaustiveness checks. It covers
the usage in switch statements and the improvements in TypeScript 5.3
with `switch(true)` narrowing. The article also compares the new
approach with the old way of using a `never` type assertion.