feat: Add new article about ray-so.vim plugin
This commit introduces a new article about the ray-so.vim plugin. The
plugin allows users to use ray.so from Neovim/Vim. The article explains
the motivation behind creating the plugin, its implementation details,
and future work. It also includes a brief introduction to ray.so and
Denops, the ecosystem used for the plugin's implementation.