feat: Add karabiner-elements config for ChatGPT app
This commit introduces a new configuration for karabiner-elements
to swap the functionality of Return and Shift + Return keys in the
Mac version of the ChatGPT app. The configuration is implemented
using karabiner.ts. The rules are activated when the Bundle
Identifier is 'com.openai.chat'. The changes are:

- When Return + Shift is pressed, a Return is sent
- When Return is pressed, Shift + Return is sent
- Command + Return sends a Return

The compiled configuration is also provided for those not using
karabiner.ts.