feat: add cloudflare redirect plugin to vite config
This commit introduces the '@ryoppippi/vite-plugin-cloudflare-redirect'
plugin to the Vite configuration. The plugin is set to generate mode
and is configured to redirect from '/cv' to 'https://cv.ryoppippi.com'
with a status of 301. The plugin and its dependencies are also added to
the package.json and the lock file is updated accordingly.