feat(svelte): upgrade Svelte to v5.0.0-next.184
This commit upgrades Svelte from v4.2.17 to v5.0.0-next.184. It also
adds a new redirect in the static/_redirects file. Additionally, it
modifies the svelte.config.js file to include a dynamicCompileOptions
function for the vitePlugin configuration. This function forces the
use of runes for user scripts due to compatibility issues with the
sveltekit-tweet and rune.