feat(ExperienceTitle.svelte): allow link to be optional
The link property in ExperienceTitle.svelte is now optional. This change
allows for instances where there may not be a relevant link to provide.
The condition for rendering the link has also been updated to check for
null values, ensuring that the link is only rendered when a valid URL is
provided.