feat(svelte.config): update asset paths and inspector config
This commit updates the `svelte.config.js` file. The `dev` constant was
removed and `isDevelopment` is used directly instead. The `assets` path
now checks if the environment is development or production and sets the
path accordingly. If it's the main branch in production, it sets the path
to `https:///vim-jp-radio.com`. Otherwise, it uses `CF_PAGES_URL`.