feat(Chip, ChipList): replace Svelte component with string for icon
In both Chip.svelte and ChipList.svelte, the `icon` prop has been
changed from a Svelte component to a string. This change simplifies
the usage of the `icon` prop and allows for more flexibility in
icon usage. The icon string is now used as a class name in a div
element.